fix(crawlers): move CSV's e2e spec into tools/crawlers/csv
The file-upload-step e2e spec was added under app/ui/e2e/, which is
exactly the "crawler-specific file living outside its crawler folder"
problem CSV's wizard/discover/summary files were already moved out of
core for. Relocated it to tools/crawlers/csv/ConfigWizard.e2e.mjs.

A colocated file can't import { test, expect } from '@playwright/test'
directly: that package is only installed under app/ui/node_modules, and
tools/crawlers/ isn't a descendant of app/ui, so Node's module resolution
can't reach it — same root cause as the Docker frontend-build's
node_modules-hoisting fix (app/api/Dockerfile), but with no equivalent
hoisting available for local/CI test runs without restructuring how the
whole project installs dependencies. Worked around it the same way
discover.js/ConfigWizard.jsx/Summary.jsx avoid hardcoding crawler
knowledge in core: the spec exports register(test, expect), and a new,
fully generic app/ui/e2e/crawler-plugin-tests.spec.js (mirroring
crawler-wizard-discovery.spec.js's discovery style) finds every
tools/crawlers/<type>/*.e2e.mjs file and registers it. The .mjs extension
is required — Playwright's loader doesn't apply Node's "detect module
syntax" auto-detection that a plain .js file here would need.

Updated tools/crawlers/CLAUDE.md's e2e-testing section to document this
pattern. Verified via a real Docker build + isolated Postgres container:
both tests still pass (2x repeat, no flakiness), and the upload/delete
round trip still actually touches disk under /data/uploads/csv-1/.
