+++ changes/feature-csv-wizard-upload-e2e-test.md
@@ -1,1 +1,2 @@
 - No user-facing change. Added end-to-end test coverage for the CSV crawler wizard's file upload step (staging files, the required-object coverage indicator, and the real upload/list/delete round trip against the server), closing a test gap that previously only covered the wizard's first step via a static render check.
+- The new test is co-located with the CSV crawler plugin (`tools/crawlers/csv/`) rather than in the core UI's e2e folder, consistent with every other crawler-specific file. Since a colocated file can't directly import Playwright's test APIs (no shared `node_modules` ancestor), it's loaded via a small generic discovery spec (`app/ui/e2e/crawler-plugin-tests.spec.js`) that contains no crawler-specific knowledge — the same pattern already used for the wizard/discover/summary plugin files.
